finished documenting methods, added README.md, set up gradle correctly, minor refactors
