+++ examples/realworld-app/build.gradle
@@ -2,6 +2,7 @@ plugins {
   id 'java'
   id 'org.springframework.boot' version '3.2.1'
   id 'io.spring.dependency-management' version '1.1.4'
+  id 'application'
 }
 
 group = 'realworldapp'
@@ -17,5 +18,8 @@ repositories {
 
 dependencies {
   implementation('org.springframework.boot:spring-boot-starter-parent:3.2.0')
+  implementation('realworldapp:rw-server')
 }
+
+mainClassName = 'rw-server/src/main/java/realworld/SpringBootApp.java'
