+++ examples/realworld-app/docker-compose.yaml
@@ -1,0 +1,29 @@
+services:
+  es:
+    image: docker.elastic.co/elasticsearch/elasticsearch:8.11.1
+    container_name: es
+    environment:
+      "discovery.type": "single-node"
+      "xpack.security.enabled": "false"
+      "xpack.security.http.ssl.enabled": "false"
+    ports:
+      - "9200:9200"
+    networks:
+      - elastic
+  kibana:
+    image: docker.elastic.co/kibana/kibana:8.7.1
+    container_name: kibana
+    environment:
+      XPACK_ENCRYPTEDSAVEDOBJECTS_ENCRYPTIONKEY: d1a66dfd-c4d3-4a0a-8290-2abcb83ab3aa
+    ports:
+      - 5601:5601
+    networks:
+      - elastic
+    deploy:
+      resources:
+        limits:
+          cpus: '2.0'
+        reservations:
+          cpus: '1.0'
+networks:
+  elastic:
