+++ examples/realworld-app/rw-database/src/main/java/realworld/document/exception/ResourceAlreadyExistsException.java
@@ -17,18 +17,11 @@
  * under the License.
  */
 
-package realworld.entity.article;
+package realworld.document.exception;
 
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
-import com.fasterxml.jackson.annotation.JsonTypeInfo.Id;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-import jakarta.validation.constraints.NotNull;
+public class ResourceAlreadyExistsException extends RuntimeException {
 
-import java.util.List;
-
-@JsonTypeName("article")
-@JsonTypeInfo(include = As.WRAPPER_OBJECT, use = Id.NAME)
-public record ArticleCreationDTO(@NotNull String title, @NotNull String description, @NotNull String body,
-                                 List<String> tagList) {
+    public ResourceAlreadyExistsException(String message) {
+        super(message);
+    }
 }
