+++ examples/realworld-app/rw-database/src/main/java/realworld/document/user/UserDTO.java
@@ -17,9 +17,23 @@
  * under the License.
  */
 
-package realworld.entity.article;
+package realworld.document.user;
 
-import java.util.List;
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
+import com.fasterxml.jackson.annotation.JsonTypeInfo.Id;
+import com.fasterxml.jackson.annotation.JsonTypeName;
 
-public record ArticlesDTO(List<ArticleForListDTO> articles, int articlesCount) {
+@JsonTypeName("user")
+@JsonTypeInfo(include = As.WRAPPER_OBJECT, use = Id.NAME)
+public record UserDTO(
+    String username,
+    String email,
+    String token,
+    String bio,
+    String image) {
+
+    public UserDTO(User user) {
+        this(user.username(), user.email(), user.token(), user.bio(), user.image());
+    }
 }
