+++ examples/realworld-app/rw-database/src/main/java/realworld/utils/ArticleIdPair.java
@@ -1,6 +1,6 @@
 package realworld.utils;
 
-import realworld.entity.article.Article;
+import realworld.document.article.Article;
 
 public record ArticleIdPair(Article article, String id) {
 }
