+++ examples/realworld-app/rw-database/src/main/java/realworld/utils/UserIdPair.java
@@ -1,6 +1,6 @@
 package realworld.utils;
 
-import realworld.entity.user.User;
+import realworld.document.user.User;
 
 public record UserIdPair(User user, String id) {
 }
