+++ examples/realworld-app/rw-rest/src/main/java/realworld/rest/TagsController.java
@@ -28,7 +28,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import realworld.db.ArticleService;
-import realworld.entity.article.TagsDTO;
+import realworld.document.article.TagsDTO;
 
 import java.io.IOException;
